requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
anthropic>=0.25.0
python-dotenv>=1.0.0
tqdm>=4.66.0
